+++ app/src/main/java/edu/coe/richmond/Item.kt
@@ -4,17 +4,17 @@ import kotlin.toString as toString1
 
 class Item {
     var name: String? = null
-    var price:Int = 0
+    var price:Double = 0.00
 
 
     constructor(){
         this.name = ""
-        this.price = 0
+        this.price = 0.00
     }
 
-    constructor(name: String?, price: Int) {
+    constructor(name: String?, price: Double) {
         this.name = name
-        this.price = price
+        this.price = price.toDouble()
     }
 
     var foodName: String?
